# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NER Smart Logistics & Accessibility Intelligence Platform",
    "description": "Backend API for the Smart Logistics & Accessibility Intelligence Platform for North-East India (SIH 2026 \u2014 Problem 26002, MDoNER). Supports field incident reporting, road network status, routing, fleet tracking, and logistics coordination for humanitarian operations.",
    "contact": {
      "name": "SIH 2026 Team"
    },
    "license": {
      "name": "Restricted \u2014 Pilot Use Only"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/logistics/vehicles": {
      "post": {
        "tags": [
          "Logistics & Fleet"
        ],
        "summary": "Create Vehicle",
        "operationId": "create_vehicle_api_v1_logistics_vehicles_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Logistics & Fleet"
        ],
        "summary": "List Vehicles",
        "operationId": "list_vehicles_api_v1_logistics_vehicles_get",
        "parameters": [
          {
            "name": "is_active",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Active"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VehicleResponse"
                  },
                  "title": "Response List Vehicles Api V1 Logistics Vehicles Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/logistics/drivers": {
      "post": {
        "tags": [
          "Logistics & Fleet"
        ],
        "summary": "Create Driver",
        "operationId": "create_driver_api_v1_logistics_drivers_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriverCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Logistics & Fleet"
        ],
        "summary": "List Drivers",
        "operationId": "list_drivers_api_v1_logistics_drivers_get",
        "parameters": [
          {
            "name": "is_active",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Active"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DriverResponse"
                  },
                  "title": "Response List Drivers Api V1 Logistics Drivers Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/logistics/commitments": {
      "post": {
        "tags": [
          "Logistics & Fleet"
        ],
        "summary": "Create Commitment",
        "operationId": "create_commitment_api_v1_logistics_commitments_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommitmentCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommitmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Logistics & Fleet"
        ],
        "summary": "List Commitments",
        "operationId": "list_commitments_api_v1_logistics_commitments_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommitmentResponse"
                  },
                  "title": "Response List Commitments Api V1 Logistics Commitments Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/logistics/trips": {
      "post": {
        "tags": [
          "Logistics & Fleet"
        ],
        "summary": "Dispatch Trip",
        "operationId": "dispatch_trip_api_v1_logistics_trips_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DispatchTripRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Logistics & Fleet"
        ],
        "summary": "List Trips",
        "operationId": "list_trips_api_v1_logistics_trips_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TripStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TripResponse"
                  },
                  "title": "Response List Trips Api V1 Logistics Trips Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/logistics/trips/{trip_id}": {
      "get": {
        "tags": [
          "Logistics & Fleet"
        ],
        "summary": "Get Trip",
        "operationId": "get_trip_api_v1_logistics_trips__trip_id__get",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Trip Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/logistics/trips/{trip_id}/transition": {
      "post": {
        "tags": [
          "Logistics & Fleet"
        ],
        "summary": "Transition Trip",
        "operationId": "transition_trip_api_v1_logistics_trips__trip_id__transition_post",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Trip Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripTransitionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/routes/{route_plan_id}": {
      "get": {
        "tags": [
          "Routing & Dispatch"
        ],
        "summary": "Get Route Plan",
        "operationId": "get_route_plan_api_v1_routes__route_plan_id__get",
        "parameters": [
          {
            "name": "route_plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Route Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutePlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CommitmentCreateRequest": {
        "properties": {
          "consignment_reference": {
            "type": "string",
            "maxLength": 64,
            "minLength": 2,
            "title": "Consignment Reference"
          },
          "cargo_category": {
            "$ref": "#/components/schemas/CargoCategory"
          },
          "priority_tier": {
            "$ref": "#/components/schemas/PriorityTier"
          },
          "consigned_weight_kg": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Consigned Weight Kg"
          },
          "consigned_quantity_units": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Consigned Quantity Units"
          },
          "origin_facility_id": {
            "type": "string",
            "format": "uuid",
            "title": "Origin Facility Id"
          },
          "destination_facility_id": {
            "type": "string",
            "format": "uuid",
            "title": "Destination Facility Id"
          },
          "required_before": {
            "type": "string",
            "format": "date-time",
            "title": "Required Before"
          },
          "consigned_volume_m3": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consigned Volume M3"
          }
        },
        "type": "object",
        "required": [
          "consignment_reference",
          "cargo_category",
          "priority_tier",
          "consigned_weight_kg",
          "consigned_quantity_units",
          "origin_facility_id",
          "destination_facility_id",
          "required_before"
        ],
        "title": "CommitmentCreateRequest"
      },
...
      "DriverCreateRequest": {
        "properties": {
          "full_name": {
            "type": "string",
            "maxLength": 128,
            "minLength": 2,
            "title": "Full Name"
          },
          "phone_e164": {
            "type": "string",
            "maxLength": 32,
            "minLength": 8,
            "title": "Phone E164"
          },
          "license_number": {
            "type": "string",
            "maxLength": 64,
            "minLength": 4,
            "title": "License Number"
          },
          "license_classes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "License Classes"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          }
        },
        "type": "object",
        "required": [
          "full_name",
          "phone_e164",
          "license_number"
        ],
        "title": "DriverCreateRequest"
      },
      "DriverResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "organization_id": {
            "type": "string",
            "format": "uuid",
            "title": "Organization Id"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "phone_e164": {
            "type": "string",
            "title": "Phone E164"
          },
          "license_number": {
            "type": "string",
            "title": "License Number"
          },
          "license_classes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "License Classes"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "organization_id",
          "user_id",
          "full_name",
          "phone_e164",
          "license_number",
          "license_classes",
          "is_active",
          "created_at"
        ],
        "title": "DriverResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TripStatus": {
        "type": "string",
        "enum": [
          "PLANNED",
          "DISPATCHED",
          "IN_TRANSIT",
          "HELD_FOR_INSPECTION",
          "DIVERTED",
          "COMPLETED",
          "CANCELLED",
          "ABORTED"
        ],
        "title": "TripStatus"
      },
...
      "TripTransitionRequest": {
        "properties": {
          "target_status": {
            "$ref": "#/components/schemas/TripStatus"
          }
        },
        "type": "object",
        "required": [
          "target_status"
        ],
        "title": "TripTransitionRequest"
      },
...
      "VehicleCreateRequest": {
        "properties": {
          "registration_number": {
            "type": "string",
            "maxLength": 32,
            "minLength": 2,
            "title": "Registration Number"
          },
          "vehicle_type": {
            "$ref": "#/components/schemas/VehicleType"
          },
          "make_model": {
            "type": "string",
            "maxLength": 128,
            "minLength": 2,
            "title": "Make Model"
          },
          "max_weight_kg": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Max Weight Kg"
          },
          "empty_weight_kg": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Empty Weight Kg"
          },
          "height_m": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Height M"
          },
          "width_m": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Width M"
          },
          "length_m": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Length M"
          },
          "axle_count": {
            "type": "integer",
            "minimum": 2.0,
            "title": "Axle Count"
          },
          "is_hazmat_capable": {
            "type": "boolean",
            "title": "Is Hazmat Capable",
            "default": false
          },
          "is_refrigerated": {
            "type": "boolean",
            "title": "Is Refrigerated",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "registration_number",
          "vehicle_type",
          "make_model",
          "max_weight_kg",
          "empty_weight_kg",
          "height_m",
          "width_m",
          "length_m",
          "axle_count"
        ],
        "title": "VehicleCreateRequest"
      },
...
      "VehicleResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "organization_id": {
            "type": "string",
            "format": "uuid",
            "title": "Organization Id"
          },
          "registration_number": {
            "type": "string",
            "title": "Registration Number"
          },
          "vehicle_type": {
            "$ref": "#/components/schemas/VehicleType"
          },
          "make_model": {
            "type": "string",
            "title": "Make Model"
          },
          "max_weight_kg": {
            "type": "number",
            "title": "Max Weight Kg"
          },
          "empty_weight_kg": {
            "type": "number",
            "title": "Empty Weight Kg"
          },
          "height_m": {
            "type": "number",
            "title": "Height M"
          },
          "width_m": {
            "type": "number",
            "title": "Width M"
          },
          "length_m": {
            "type": "number",
            "title": "Length M"
          },
          "axle_count": {
            "type": "integer",
            "title": "Axle Count"
          },
          "is_hazmat_capable": {
            "type": "boolean",
            "title": "Is Hazmat Capable"
          },
          "is_refrigerated": {
            "type": "boolean",
            "title": "Is Refrigerated"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "organization_id",
          "registration_number",
          "vehicle_type",
          "make_model",
          "max_weight_kg",
          "empty_weight_kg",
          "height_m",
          "width_m",
          "length_m",
          "axle_count",
          "is_hazmat_capable",
          "is_refrigerated",
          "is_active",
          "created_at"
        ],
        "title": "VehicleResponse"
      },
      "VehicleType": {
        "type": "string",
        "enum": [
          "TRUCK_HEAVY",
          "TRUCK_MEDIUM",
          "VAN_LIGHT",
          "FOUR_WHEEL_DRIVE",
          "AMBULANCE_RESCUE",
          "TWO_WHEEL_SPECIAL"
        ],
        "title": "VehicleType"
      }
    }
  }
}